requirements.txt
openai==0.27.4
python-dotenv==1.0.0
requests==2.28.1
